backend/requirements.txt
fastapi
uvicorn[standard]
langchain
psycopg2-binary
pymongo
pydantic
sqlalchemy
pymilvus
python-dotenv
loguru
httpx 